Create Day20 : Sorting
Task 
Given an array, , of size  distinct elements, sort the array in ascending order using the Bubble Sort algorithm above. Once sorted, print the following  lines:

Array is sorted in numSwaps swaps. 
where  is the number of swaps that took place.
First Element: firstElement 
where  is the first element in the sorted array.
Last Element: lastElement 
where  is the last element in the sorted array.
Hint: To complete this challenge, you will need to add a variable that keeps a running tally of all swaps that occur during execution.