Added description to File
Task 
A prime is a natural number greater than  that has no positive divisors other than 1 and itself. Given a number, , determine and print whether it's Prime or Not Prime.

Note: If possible, try to come up with a O(sqrt(n)) primality algorithm, or see what sort of optimizations you come up with for an  algorithm. Be sure to check out the Editorial after submitting your code!